requirements.txt
streamlit
PyPDF2
python-docx
fpdf
pandas
beautifulsoup4
lxml
